+++ Changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Long title infos are abbreviated with '...'
 * Status bar now shows the correct version
+* Volume is no longer reset when playing the next song
 
 ### Changed